This project is running on BlackPill with STM32F401CC6 !
The code collects analog values from 10 ADC channels in scan mode and uses the DMA
to update the raw values in array adc1RowValuesFromDma

The ADC_Handling.c code uses a second array ( adc1ReadRowValues ) to average the reading.
The reading is currently smoothed with a window of 8 samples.

The ADC is triggered by software at 1,s intervals (in main.c )

The UART_Handling.c code reads back the smoothed data and sends a message to the listening host.

Sending is called from main.c and currently is handled every 50ms